Generalize the plugin system.
This change pulls the plugin system out of App and makes it a generic
system that can extend any class with a plugin system. This was done
in anticipation of a new WebSocket server class, which will also have
App-style plugins, but with a different API, and loaded from a different
place in the filesystem.

This change requires code changes to plugins and the classes that
are extended by them:

* Strelka::App::Plugin is now Strelka::Plugin
* Strelka::App::Plugins is now Strelka::PluginLoader

--HG--
rename : spec/strelka/app/plugins_spec.rb => spec/strelka/plugins_spec.rb